executor-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pywinrm==0.4.3
paramiko==3.4.0
jinja2==3.1.2
pydantic==2.5.0
aiohttp==3.9.1
httpx>=0.24.0
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
flower==2.0.1
